requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
distlib==0.3.2
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
djangorestframework==3.12.4
environ==1.0
filelock==3.0.12
idna==3.2
pipenv==2021.5.29
platformdirs==2.3.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.8.0
virtualenv-clone==0.5.7

